Corrected alloc/2 that did not work in R13B.
alloc/2 used to call the driver that would allocate a driver binary.
In R13B, however, a binary smaller than 65 bytes will be
automatically converted to a heap binary.

We work around the problem by always allocating a binary with
a size of at least 65 bytes. We do it directly in Erlang code
without calling the driver (there was never really any need
to use the driver just to allocate a new binary).