sdl_video:wm_setIcon/2 with a mask now works.
The Erlang code in sdl_video:wm_setIcon/2 had type errors that
made it crash if the second mask argument was something other
than 'null'.

The driver code would refuse to use a bit mask of size 32 x 32
or greater because it insisted on copying the mask into a
stack-allocated array. I see no need to copy the mask, so
now it is passed directly to SDL_WM_SetIcon(), allowing it
to have any size.